# Add `CachedManifestLoader`