--- a/Sources/TuistLoader/Loaders/CachedManifestLoader.swift
+++ b/Sources/TuistLoader/Loaders/CachedManifestLoader.swift
@@ -15,28 +15,37 @@ public class CachedManifestLoader: ManifestLoading {
     private let helpersDirectoryLocator: HelpersDirectoryLocating
     private let cacheDirectory: AbsolutePath
     private let fileHandler: FileHandling
+    private let environment: Environmenting
+    private let tuistVersion: String
     private let decoder: JSONDecoder = JSONDecoder()
     private let encoder: JSONEncoder = JSONEncoder()
     private var helpersCache: [AbsolutePath: String?] = [:]
 
     public convenience init(manifestLoader: ManifestLoading = ManifestLoader()) {
+        let environment = TuistSupport.Environment.shared
         self.init(manifestLoader: manifestLoader,
                   projectDescriptionHelpersHasher: ProjectDescriptionHelpersHasher(),
                   helpersDirectoryLocator: HelpersDirectoryLocator(),
-                  cacheDirectory: Environment.shared.cacheDirectory.appending(component: "Manifests"),
-                  fileHandler: FileHandler.shared)
+                  cacheDirectory: environment.cacheDirectory.appending(component: "Manifests"),
+                  fileHandler: FileHandler.shared,
+                  environment: environment,
+                  tuistVersion: Constants.version)
     }
 
     init(manifestLoader: ManifestLoading,
          projectDescriptionHelpersHasher: ProjectDescriptionHelpersHashing,
          helpersDirectoryLocator: HelpersDirectoryLocating,
          cacheDirectory: AbsolutePath,
-         fileHandler: FileHandling) {
+         fileHandler: FileHandling,
+         environment: Environmenting,
+         tuistVersion: String) {
         self.manifestLoader = manifestLoader
         self.projectDescriptionHelpersHasher = projectDescriptionHelpersHasher
         self.helpersDirectoryLocator = helpersDirectoryLocator
         self.cacheDirectory = cacheDirectory
         self.fileHandler = fileHandler
+        self.environment = environment
+        self.tuistVersion = tuistVersion
     }
 
     public func loadConfig(at path: AbsolutePath) throws -> Config {
@@ -77,13 +86,18 @@ public class CachedManifestLoader: ManifestLoading {
             throw ManifestLoaderError.manifestNotFound(manifest, path)
         }
 
-        let manifestHash = try calculateManifestHash(for: manifest, at: manifestPath)
-        let helpersHash = try calculateHelpersCache(at: path)
+        let calculatedHashes = try? calculateHashes(path: path,
+                                                    manifestPath: manifestPath,
+                                                    manifest: manifest)
+
+        guard let hashes = calculatedHashes else {
+            logger.warning("Unable to calculate manifest hash at path: \(path)")
+            return try loader()
+        }
 
         let cachedManifestPath = cachedPath(for: manifestPath)
         if let cached: T = loadCachedManifest(at: cachedManifestPath,
-                                              manifestHash: manifestHash,
-                                              helpersHash: helpersHash) {
+                                              hashes: hashes) {
             return cached
         }
 
@@ -91,11 +105,22 @@ public class CachedManifestLoader: ManifestLoading {
 
         try cacheManifest(manifest: manifest,
                           loadedManifest: loadedManifest,
-                          manifestHash: manifestHash,
-                          helpersHash: helpersHash,
+                          hashes: hashes,
                           to: cachedManifestPath)
 
         return loadedManifest
+    }
+
+    private func calculateHashes(path: AbsolutePath,
+                                 manifestPath: AbsolutePath,
+                                 manifest: Manifest) throws -> Hashes {
+        let manifestHash = try calculateManifestHash(for: manifest, at: manifestPath)
+        let helpersHash = try calculateHelpersHash(at: path)
+        let environmentHash = calculateEnvironmentHash()
+
+        return Hashes(manifestHash: manifestHash,
+                      helpersHash: helpersHash,
+                      environmentHash: environmentHash)
     }
 
     private func calculateManifestHash(for manifest: Manifest, at path: AbsolutePath) throws -> Data {
@@ -105,7 +130,7 @@ public class CachedManifestLoader: ManifestLoading {
         return hash
     }
 
-    private func calculateHelpersCache(at path: AbsolutePath) throws -> String? {
+    private func calculateHelpersHash(at path: AbsolutePath) throws -> String? {
         guard let helpersDirectory = helpersDirectoryLocator.locate(at: path) else {
             return nil
         }
@@ -120,16 +145,23 @@ public class CachedManifestLoader: ManifestLoading {
         return hash
     }
 
+    private func calculateEnvironmentHash() -> String? {
+        let tuistEnvVariables = environment.tuistVariables.map { "\($0.key)=\($0.value)" }.sorted()
+        guard !tuistEnvVariables.isEmpty else {
+            return nil
+        }
+        return tuistEnvVariables.joined(separator: "-").md5
+    }
+
     private func cachedPath(for manifestPath: AbsolutePath) -> AbsolutePath {
         let pathHash = manifestPath.pathString.md5
-        let cacheVersion = CachedManifest.currentVersion.description
+        let cacheVersion = CachedManifest.currentCacheVersion.description
         let fileName = [cacheVersion, pathHash].joined(separator: ".")
         return cacheDirectory.appending(component: fileName)
     }
 
     private func loadCachedManifest<T: Decodable>(at cachedManifestPath: AbsolutePath,
-                                                  manifestHash: Data,
-                                                  helpersHash: String?) -> T? {
+                                                  hashes: Hashes) -> T? {
         guard fileHandler.exists(cachedManifestPath) else {
             return nil
         }
@@ -142,9 +174,9 @@ public class CachedManifestLoader: ManifestLoading {
             return nil
         }
 
-        guard cachedManifest.version == CachedManifest.currentVersion,
-            cachedManifest.helpersHash == helpersHash,
-            cachedManifest.manifestHash == manifestHash else {
+        guard cachedManifest.cacheVersion == CachedManifest.currentCacheVersion,
+            cachedManifest.tuistVersion == tuistVersion,
+            cachedManifest.hashes == hashes else {
             return nil
         }
 
@@ -153,11 +185,10 @@ public class CachedManifestLoader: ManifestLoading {
 
     private func cacheManifest<T: Encodable>(manifest: Manifest,
                                              loadedManifest: T,
-                                             manifestHash: Data,
-                                             helpersHash: String?,
+                                             hashes: Hashes,
                                              to cachedManifestPath: AbsolutePath) throws {
-        let cachedManifest = CachedManifest(manifestHash: manifestHash,
-                                            helpersHash: helpersHash,
+        let cachedManifest = CachedManifest(tuistVersion: tuistVersion,
+                                            hashes: hashes,
                                             manifest: try encoder.encode(loadedManifest))
 
         let cachedManifestData = try encoder.encode(cachedManifest)
@@ -174,10 +205,18 @@ public class CachedManifestLoader: ManifestLoading {
     }
 }
 
-private struct CachedManifest: Codable {
-    static let currentVersion = 1
-    var version: Int = currentVersion
+private struct Hashes: Equatable, Codable {
     var manifestHash: Data
     var helpersHash: String?
+    var environmentHash: String?
+}
+
+private struct CachedManifest: Codable {
+    // Note: please bump the version in case the cache structure is modifed
+    // this ensures older cache versions are not loaded using this structure
+    static let currentCacheVersion = 1
+    var cacheVersion: Int = currentCacheVersion
+    var tuistVersion: String
+    var hashes: Hashes
     var manifest: Data
 }