… CHANGELOG.md
… Tests/TuistLoaderTests/ProjectDescriptionHelpers/Mocks/MockProjectDescriptionHelpersHasher.swift