Part of: https://github.com/tuist/tuist/issues/1042

### Short description 📝

One of the areas that contribute to slow generation times is the manifest loading process (see #1042). Loading a single Swift manifest involves compiling and running that manifest along with any helper files if present. While an individual manifest doesn't take a huge amount of time, it adds up over a large workspace with several manifests.

### Solution 📦

Caching the Swift manifest in json format can significantly speed up the loading process. This can be achieved while keeping the Swift manifests the source of truth via:

- Adding a new caching implementation of the manifest loader
- It works by wrapping the existing manifest loader and adds a caching layer ontop of it
- It attempts to first load a cached JSON representation of the manifest from `~/.tuist/Cache/Manifests` and then falls back to loading the manifest using the default implementation
- The hashes of the manifest content as well the helpers (`ProjectDescriptionHelpers`) are included with the cached manifest
- This allows comparing those hashes during the loading process and ensure the cache is updated from source when needed
- Additionally a version number is also included in the cached data should the caching format change in the future
- Overall this helps speed up the manifest loading process significantly as loading JSON from disk is much faster than compiling and running Swift manifest
- This approach ensures Swift manifests continue to be the source of truth while offering an optimization for the cases where the manifest doesn't change
- Caching can be enabled / disabled via `TUIST_CACHE_MANIFESTS=1` or `TUIST_CACHE_MANIFESTS=0`
- By default its enabled however this can be changed if we find any issues with this feature

### Implementation 👩‍💻👨‍💻

- [x] Add `CachedManifestLoader`
- [x] Benchmark generation times
- [x] Profile / benchmark manifest loading times in instruments using sign posts
- [x] Update change log

### Test Plan 🛠

Modifying manifests:
- Run `tuist generate` on any of the fixtures
- Inspect the generation time
- Re-run `tuist generate`
- Inspect the generation time, which should be faster
  - Note: most gains will be seen in larger projects with several manifests
- Modify any the project manifest (e.g. update the name)
- Re-run `tuist generate`
- Verify the updates are reflected in the generated project (i.e. the cache isn't incorrectly being used)

Modifying helpers:
- Repeat the test using a `fixture/ios_app_with_helpers`
- This time modify `Tuist/ProjectDescriptionHelpers`
- Re-run `tuist generate`
- Verify the updates are reflected in the generated project

Disabling caching:
- Remove the contents of `~/.tuist/Cache/Manifests`
- Run `TUIST_CACHE_MANIFESTS=0 tuist generate` on any of the fixtures
- Inspect the generation time
- Verify no cached manifests are added to `~/.tuist/Cache/Manifests`

### Benchmark Results ⏱

With this approach there is a small increase to initial generation time (cold run) **~1.5%**, however subsequent generation times (warm run) is **~ 50%** faster!

- **New** is this PR
- **Old** is latest release (1.8.0)

| Fixture         | New    | Old  | Delta    |
| --------------- | ------ | ---- | -------- |
| 20_projects _(cold)_ | 10.14s | 10.00s | ⬆︎ +0.15s +1.48% |
| 20_projects _(warm)_ | 4.63s | 9.79s | ⬇︎ -5.16s -52.72% |
| 50_projects _(cold)_ | 23.90s | 24.12s | ⬇︎ -0.22s -0.92% |
| 50_projects _(warm)_ | 11.65s | 23.90s | ⬇︎ -12.26s -51.27% |
| ios_app_with_helpers _(cold)_ | 2.26s | 2.23s | ⬆︎ +0.03s +1.15% |
| ios_app_with_helpers _(warm)_ | 1.12s | 5.06s | ⬇︎ -3.93s -77.80% |
| ios_app_with_tests _(cold)_ | 0.85s | 0.86s | ≈ |
| ios_app_with_tests _(warm)_ | 0.66s | 0.86s | ⬇︎ -0.20s -23.63% |
| ios_app_with_custom_workspace _(cold)_ | 1.75s | 1.73s | ⬆︎ +0.02s +1.29% |
| ios_app_with_custom_workspace _(warm)_ | 0.99s | 1.74s | ⬇︎ -0.75s -42.95% |

Here are the steps taken to get those measurements, some were generated fixtures while others were existing ones within `fixtures/`

- Create a large fixture:

```sh
cd tools/fixturegen
swift run fixturegen --projects 50

FIXTURE_PATH="$(pwd)/Fixture"
```

- Benchmark:

```sh
swift build -c release --product ProjectDescription
swift build -c release --product tuist

LOCAL_TUIST="$(swift build -c release --show-bin-path)/tuist"
REFERENCE_TUIST="$(which tuist)"

cd tools/tuistbench
swift run tuistbench -b "$LOCAL_TUIST" -r "$REFERENCE_TUIST" -f "$FIXTURE_PATH" --format markdown
```

### Profiling Results

For a workspace with 50 projects and a Tuist config file

**Caching disabled**
| Component       |       Count       |       Duration      |       Min Duration      |       Avg Duration   |      Std Dev Duration         |        Max Duration        |      
| --------------- | -------- | ---------- | ----------- |  ------------- | ---------- | -------- |
| ManifestLoader.loadManifest | 103 | 36.52 s | 281.19 ms | 354.53 ms | 88.05 ms | 1.03 s | 

**Caching enabled**

When there is no cache:
| Component       |       Count       |       Duration      |       Min Duration      |       Avg Duration   |      Std Dev Duration         |        Max Duration        |      
| --------------- | -------- | ---------- | ----------- |  ------------- | ---------- | -------- |
| CachedManifestLoader.load    |   103 |  25.63 s  |    601.28 µs | 248.84 ms    |   332.86 ms   |  2.65 s |  
| CachedManifestLoader.cacheManifests   |   52 |   252.14 ms    | 3.14 ms |    4.85 ms  |  1.67 ms   |   9.78 ms    |   

When cache is warm:

| Component       |       Count       |       Duration      |       Min Duration      |       Avg Duration   |      Std Dev Duration         |        Max Duration        |      
| --------------- | -------- | ---------- | ----------- |  ------------- | ---------- | -------- |
| CachedManifestLoader.load | 103 | 451.41 ms | 464.56 µs | 4.38 ms | 8.62 ms | 83.16 ms | 

**This PR improves the manifest loading process by ~98%*  🎉

---

You'll notice the count is 103 even though the workspace has 50 projects and 1 Config files. Looks like we're redundantly loading `Config.swift` once per manifest 😱 - this can be fixed in a separate PR.

![Screen Shot 2020-05-27 at 6 32 11 PM](https://user-images.githubusercontent.com/11914919/83053287-68872c00-a048-11ea-9602-b06cddd02400.png)